backend/requirements.txt
# Django & Django REST Framework
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# Database
mysqlclient==2.2.1

# Authentication & Security
PyJWT==2.8.0

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.0.0

# API Documentation
drf-spectacular==0.27.0

# Utilities
Pillow==10.2.0
markdown==3.5.1

# Development
ipython==8.20.0

# Production Server
gunicorn==21.2.0


